+++ Models/Entities/BaseIdentityEntity.cs
@@ -9,13 +9,26 @@ namespace Models.Entities;
 /// </summary>
 /// <remarks>
 /// Inherits from <see cref="IdentityUser{TKey}"/> with <see langword="long"/> as the key
-/// type. Inherit from <see cref="BaseIdentityEntity"/> to add standardized auditing properties
-/// to your user domain entities. The <see cref="CreatedAt"/> property is automatically
+/// type.<br/>
+/// Inherit from <see cref="BaseIdentityEntity"/> to add standardized auditing properties
+/// to your domain entities. The <see cref="CreatedAt"/> property is automatically
 /// initialized to the current UTC date and time when the entity is instantiated. The <see
 /// cref="DeletedAt"/> property can be set to indicate when the entity was deleted, or left
 /// <see langword="null"/> if the entity is active.
 /// </remarks>
 public abstract class BaseIdentityEntity : IdentityUser<long> {
+	/// <summary>
+	/// Gets or sets the globally unique identifier for the entity.
+	/// </summary>
+	/// <value>
+	/// A long representing the primary identifier in external systems.
+	/// </value>
+	[Range(1, long.MaxValue, ErrorMessage = "GlobalId must be a positive value.")]
+	[Display(Name = "GlobalId")]
+	public long? GlobalId {
+		get; set;
+	}
+
 	/// <summary>
 	/// Gets or sets the date and time when the entity was created.
 	/// </summary>
@@ -23,7 +36,7 @@ public abstract class BaseIdentityEntity : IdentityUser<long> {
 	[DisplayFormat(DataFormatString = "{0:dd-MM-yyyy HH:mm}", ApplyFormatInEditMode = true)]
 	[Display(Name = "CreatedAt")]
 	public DateTime CreatedAt {
-		get; private set;
+		get; set;
 	} = DateTime.UtcNow;
 
 	/// <summary>