+++ Models/Entities/DTO/CoffeeDTO.cs
@@ -1,0 +1,8 @@
+﻿namespace Models.Entities.DTO;
+
+/// <summary>
+/// Data Transfer Object representing a simplified version of a Coffee, 
+/// used for transporting data across application layers.
+/// </summary>
+public class CoffeeDTO : Coffee {
+}