+++ Models/Migrations/20260124162619_InitialCreate.cs
@@ -236,9 +236,9 @@ namespace Models.Migrations
                     PaymentDetailId = table.Column<long>(type: "bigint", nullable: false)
                         .Annotation("SqlServer:Identity", "1, 1"),
                     CoffeeUserId = table.Column<long>(type: "bigint", nullable: false),
-                    LastFour = table.Column<string>(type: "nvarchar(4)", maxLength: 4, nullable: false),
+                    LastFour = table.Column<int>(type: "int", nullable: false),
                     ExpiryDate = table.Column<DateTime>(type: "datetime2", nullable: false),
-                    GatewayToken = table.Column<string>(type: "nvarchar(255)", maxLength: 255, nullable: false),
+                    GatewayToken = table.Column<string>(type: "nvarchar(max)", nullable: false),
                     CreatedAt = table.Column<DateTime>(type: "datetime2", nullable: false),
                     DeletedAt = table.Column<DateTime>(type: "datetime2", nullable: true)
                 },