+++ Web/Controllers/HomeController.cs
@@ -1,4 +1,6 @@
 using Microsoft.AspNetCore.Mvc;
+using System.Diagnostics;
+using Web.Models;
 
 namespace Web.Controllers;
 
@@ -31,10 +33,10 @@ public class HomeController : Controller {
 		return View();
 	}
 
-	//[ResponseCache(Duration = 0, Location = ResponseCacheLocation.None, NoStore = true)]
-	//public IActionResult Error() {
-	//	return View(new ErrorViewModel { RequestId = Activity.Current?.Id ?? HttpContext.TraceIdentifier });
-	//}
+	[ResponseCache(Duration = 0, Location = ResponseCacheLocation.None, NoStore = true)]
+	public IActionResult Error() {
+		return View(new ErrorViewModel { RequestId = Activity.Current?.Id ?? HttpContext.TraceIdentifier });
+	}
 
 	public IActionResult ManageCustomers() {
 		return View();